faas_get_cache/requirements.txt
Flask==1.1.2
google-cloud-firestore==2.0.1
